# Conflicts:
#	src/main/java/ru/yandex/practicum/filmorate/controllers/FilmController.java
#	src/main/java/ru/yandex/practicum/filmorate/controllers/UserController.java
#	src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
#	src/main/java/ru/yandex/practicum/filmorate/service/LikeService.java
#	src/main/java/ru/yandex/practicum/filmorate/service/ReviewService.java
#	src/main/java/ru/yandex/practicum/filmorate/service/UserFeedService.java
#	src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
#	src/main/java/ru/yandex/practicum/filmorate/storage/like/LikeDbStorage.java
#	src/main/java/ru/yandex/practicum/filmorate/storage/like/LikeStorage.java
#	src/main/resources/schema.sql
#	src/test/java/ru/yandex/practicum/filmorate/service/FeedServiceTest.java

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -37,9 +37,10 @@ public class Film {
     @NotNull
     private Mpa mpa;
 
-    private int rate;
+    private double rate;
 
     private List<Genre> genres;
 
     private List<Director> directors;
+
 }
